refactor: Use IEnumerable<char> for hot path.
The HasWord and HasPrefix wanted a string but because strings are
immutable, it required an allocation of a new string for every check I
wanted to do. Instead, I changed the API to accept IEnumerable<char>,
which string implements so that those methods will accept strings,
char[]s, and IEnumerable<char>s. That way I don't have to allocate a
string 60 times a second possibly.